refactor(resolver): refactor code to allow using strategy pattern
Refs #2744
