+++ src/index.js
@@ -1,7 +1,8 @@
 import Url from 'url';
 
 import Http, { makeHttp, serializeRes, serializeHeaders } from './http/index.js';
-import Resolver, { clearCache } from './resolver.js';
+import Resolver from './resolver/index.js';
+import { clearCache } from './resolver/strategies/openapi-2--3-0.js';
 import resolveSubtree from './subtree-resolver/index.js';
 import { makeApisTagOperation } from './interfaces.js';
 import { execute, buildRequest, baseUrl } from './execute/index.js';
